refactor: moved JwtPayload and EnvConfig types to shared-common
