+++ services/account-service/src/configs/env.config.ts
@@ -1,10 +1,8 @@
 import * as dotenv from 'dotenv';
 import path from 'path';
-import { EnvConfig } from '../types';
+import { EnvConfig } from 'shared-common/types';
 
 dotenv.config({ path: path.resolve(__dirname, '../../.env') });
-
-// console.log("Environment Variables: ", process.env);
 
 const envConfig: EnvConfig = {
   port: parseInt(process.env.PORT || '3002', 10),
