--- services/account-service/src/types/config.types.ts
@@ -1,11 +1,0 @@
-export interface EnvConfig {
-  port: number;
-  mongoURI: string;
-  jwtSecret: string;
-  logLevel: string;
-  nodeENV: string;
-  userService: string;
-  transactionService: string;
-  paymentService: string;
-  notificationService: string;
-}
