+++ services/account-service/src/types/index.ts
@@ -1,5 +1,3 @@
-export * from './config.types';
 export * from './schema.types';
 export * from './wallet.types';
 export * from './payment.types';
-// export * from './util.types';
