--- services/kyc-service/src/types/config.types.ts
@@ -1,9 +1,0 @@
-export interface EnvConfig {
-  port: number;
-  mongoURI: string;
-  jwtSecret: string;
-  logLevel: string;
-  nodeENV: string;
-  userService: string;
-  notificationService: string;
-}
