+++ services/kyc-service/src/types/index.ts
@@ -1,4 +1,2 @@
-export * from './config.types';
 export * from './kyc.types';
 export * from './schema.types';
-export * from './util.types';
