+++ services/notification-service/src/controllers/notif.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express';
 import { NotificationService } from '../services';
 import { logger } from '../utils';
-import { JwtPayload } from '../types';
+import { JwtPayload } from 'shared-common';
 
 export class NotificationController {
   constructor(private notifService: NotificationService) {}
