--- services/notification-service/src/types/config.types.ts
@@ -1,10 +1,0 @@
-export interface EnvConfig {
-  port: number;
-  mongoURI: string;
-  jwtSecret: string;
-  logLevel: string;
-  nodeENV: string;
-  serviceSecret: string;
-  mailHogUrl: string;
-  userService: string;
-}
