+++ services/notification-service/src/types/index.ts
@@ -1,3 +1,1 @@
-export * from './config.types';
-export * from './util.types';
 export * from './email.types';
