+++ services/payment-service/package.json
@@ -15,6 +15,7 @@
   "license": "ISC",
   "description": "Microservice for payment related operations",
   "dependencies": {
-    "stripe": "^16.8.0"
+    "stripe": "^16.8.0",
+    "shared-common": "^1.0.0"
   }
 }
