+++ services/payment-service/src/container.ts
@@ -3,7 +3,7 @@ import { StripeController } from './controllers';
 import { envConfig } from './configs';
 import Stripe from 'stripe';
 
-const stripe = new Stripe(envConfig.stripeSecret);
+const stripe = new Stripe(envConfig.stripeSecret as string);
 
 const stripeService = new StripeService(stripe);
 const stripeController = new StripeController(stripeService);
