+++ services/payment-service/src/types/index.ts
@@ -1,5 +1,3 @@
-export * from './config.types';
 export * from './schema.types';
-export * from './util.types';
 export * from './customer.types';
 export * from './payment.types';
