+++ services/transaction-service/src/types/index.ts
@@ -1,3 +1,2 @@
-export * from './config.types';
 export * from './schema.types';
-export * from './util.types';
+export * from './transaction.types';
