--- services/transaction-service/src/types/util.types.ts
@@ -1,8 +1,0 @@
-export interface JwtPayload {
-  sub: string;
-  email: string;
-  roles: string[];
-  status: string;
-  kycStatus: string;
-  exp: number;
-}
