+++ services/user-service/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './config.types';
 export * from './role.types';
 export * from './schema.types';
 export * from './user.types';
