+++ shared/shared-common/types/index.ts
@@ -1,1 +1,2 @@
 export * from './util.types';
+export * from './config.types';
